requirements.txt
moviepy
requests 
python-dotenv
pydub